requirements.txt
# Runtime target: Python 3.10+
torch>=2.0.0
torchvision>=0.15.0
rembg>=2.0.0
trimesh>=4.0.0
numpy>=1.24.0
Pillow>=10.0.0

# Dev
pytest>=7.4.0
